+++ bindings/python/src/encoding.rs
@@ -114,7 +114,7 @@ impl PyEncoding {
     /// Returns:
     ///     :obj:`List[int]`: The list of IDs
     #[getter]
-    fn get_ids(&self) -> Vec<u32> {
+    fn get_ids(&self) -> Vec<u64> {
         self.encoding.get_ids().to_vec()
     }
 
@@ -146,7 +146,7 @@ impl PyEncoding {
     /// Returns:
     ///     A :obj:`List` of :obj:`Optional[int]`: A list of optional word index.
     #[getter]
-    fn get_words(&self, py: Python<'_>) -> PyResult<Vec<Option<u32>>> {
+    fn get_words(&self, py: Python<'_>) -> PyResult<Vec<Option<usize>>> {
         deprecation_warning(
             py,
             "0.9.4",
@@ -168,7 +168,7 @@ impl PyEncoding {
     /// Returns:
     ///     A :obj:`List` of :obj:`Optional[int]`: A list of optional word index.
     #[getter]
-    fn get_word_ids(&self) -> Vec<Option<u32>> {
+    fn get_word_ids(&self) -> Vec<Option<usize>> {
         self.encoding.get_word_ids().to_vec()
     }
 
@@ -193,7 +193,7 @@ impl PyEncoding {
     /// Returns:
     ///     :obj:`List[int]`: The list of type ids
     #[getter]
-    fn get_type_ids(&self) -> Vec<u32> {
+    fn get_type_ids(&self) -> Vec<u64> {
         self.encoding.get_type_ids().to_vec()
     }
 
@@ -216,7 +216,7 @@ impl PyEncoding {
     /// Returns:
     ///     :obj:`List[int]`: The special tokens mask
     #[getter]
-    fn get_special_tokens_mask(&self) -> Vec<u32> {
+    fn get_special_tokens_mask(&self) -> Vec<u64> {
         self.encoding.get_special_tokens_mask().to_vec()
     }
 
@@ -229,7 +229,7 @@ impl PyEncoding {
     /// Returns:
     ///    :obj:`List[int]`: The attention mask
     #[getter]
-    fn get_attention_mask(&self) -> Vec<u32> {
+    fn get_attention_mask(&self) -> Vec<u64> {
         self.encoding.get_attention_mask().to_vec()
     }
 
@@ -265,7 +265,7 @@ impl PyEncoding {
     ///     :obj:`Tuple[int, int]`: The range of tokens: :obj:`(first, last + 1)`
     #[pyo3(signature = (word_index, sequence_index = 0))]
     #[pyo3(text_signature = "(self, word_index, sequence_index=0)")]
-    fn word_to_tokens(&self, word_index: u32, sequence_index: usize) -> Option<(usize, usize)> {
+    fn word_to_tokens(&self, word_index: usize, sequence_index: usize) -> Option<(usize, usize)> {
         self.encoding.word_to_tokens(word_index, sequence_index)
     }
 
@@ -281,7 +281,7 @@ impl PyEncoding {
     ///     :obj:`Tuple[int, int]`: The range of characters (span) :obj:`(first, last + 1)`
     #[pyo3(signature = (word_index, sequence_index = 0))]
     #[pyo3(text_signature = "(self, word_index, sequence_index=0)")]
-    fn word_to_chars(&self, word_index: u32, sequence_index: usize) -> Option<Offsets> {
+    fn word_to_chars(&self, word_index: usize, sequence_index: usize) -> Option<Offsets> {
         self.encoding.word_to_chars(word_index, sequence_index)
     }
 
@@ -332,7 +332,7 @@ impl PyEncoding {
     /// Returns:
     ///     :obj:`int`: The index of the word in the relevant input sequence.
     #[pyo3(text_signature = "(self, token_index)")]
-    fn token_to_word(&self, token_index: usize) -> Option<u32> {
+    fn token_to_word(&self, token_index: usize) -> Option<usize> {
         let (_, word_idx) = self.encoding.token_to_word(token_index)?;
         Some(word_idx)
     }
@@ -365,7 +365,7 @@ impl PyEncoding {
     ///     :obj:`int`: The index of the word that contains this char in the input sequence
     #[pyo3(signature = (char_pos, sequence_index = 0))]
     #[pyo3(text_signature = "(self, char_pos, sequence_index=0)")]
-    fn char_to_word(&self, char_pos: usize, sequence_index: usize) -> Option<u32> {
+    fn char_to_word(&self, char_pos: usize, sequence_index: usize) -> Option<usize> {
         self.encoding.char_to_word(char_pos, sequence_index)
     }
